+++ .gitignore
@@ -91,3 +91,4 @@ iOSInjectionProject/
 .DS_Store
 *.xcodeproj
 Swish/.swiftpm
+base/AppIcon.xcassets/
