+++ Swish/Sources/AppIconLib/AppIcon.Scale.swift
@@ -1,0 +1,7 @@
+extension AppIcon {
+  enum Scale: String {
+    case one = ""
+    case two = "@2x"
+    case three = "@3x"
+  }
+}
