+++ Swish/Sources/AppIconLib/AppIcon.swift
@@ -2,7 +2,7 @@ import Foundation
 import Fork
 import Sh
 
-public struct XCAssets {
+public struct AppIcon {
 
   let inputSVG: String, outputDir: String
   public init(inputSVG: String, outputDir: String) {
@@ -13,7 +13,7 @@ public struct XCAssets {
   public func render() async throws {
     try await sh(.terminal, "mkdir -p \(outputDir)/AppIcon.xcassets/AppIcon.appiconset")
 
-    try await AppiconSize.allCases.asyncForEach { size in
+    try await AppIcon.Size.allCases.asyncForEach { size in
       let pngPath = "\(outputDir)/AppIcon.xcassets/AppIcon.appiconset/\(filename(for: size))"
       try await sh(.terminal, "inkscape -o \(pngPath) -w \(size.edge) -h \(size.edge) \(inputSVG)")
     }
@@ -22,7 +22,7 @@ public struct XCAssets {
     try XCAssetsJSON.appiconsetContents.data(using: .utf8)!.write(to: URL(fileURLWithPath: "\(outputDir)/AppIcon.xcassets/AppIcon.appiconset/Contents.json"))
   }
 
-  func filename(for size: AppiconSize) -> String {
+  func filename(for size: AppIcon.Size) -> String {
     switch size.data.scale {
     case .one:
       return "AppIcon-\(size.data.pixelWidth).png"
