--- Swish/Sources/AppIconLib/AppiconScale.swift
@@ -1,5 +1,0 @@
-enum AppiconScale: String {
-  case one = ""
-  case two = "@2x"
-  case three = "@3x"
-}
